feat(input): extract `InputPlugin` into its own library.
