+++ examples/gaia-project/logic/mod.rs
@@ -1,1 +1,2 @@
+pub mod input;
 pub mod power;
