+++ examples/gaia-project/view/mod.rs
@@ -1,11 +1,12 @@
+use bevy::prelude::*;
+
 use crate::asset_library::AssetLibrary;
 use crate::logic::power::PowerCycleTracker;
 use crate::GameState;
-use bevy::prelude::*;
 
 mod power;
 
-use power::{power_view, PowerEvent, PowerView, PowerViewState};
+use power::{power_view, PowerView, PowerViewState};
 
 pub struct ViewPlugin;
 
@@ -15,12 +16,7 @@ impl Plugin for ViewPlugin {
             tracker: PowerCycleTracker::new(2, 4, 0, 0),
         })
         .add_system_set(SystemSet::on_enter(GameState::Running).with_system(init))
-        .add_system_set(
-            SystemSet::on_update(GameState::Running)
-                .with_system(power_view)
-                .with_system(input_monitor),
-        )
-        .add_event::<PowerEvent>();
+        .add_system_set(SystemSet::on_update(GameState::Running).with_system(power_view));
     }
 }
 
@@ -65,45 +61,4 @@ fn init(mut commands: Commands, asset_library: Res<AssetLibrary>) {
             ..Default::default()
         })
         .insert(PowerView);
-}
-
-// TODO: This should eventually become it's own plugin
-fn input_monitor(input: Res<Input<KeyCode>>, mut events: EventWriter<PowerEvent>) {
-    for &(key, amount) in [
-        (KeyCode::Key1, 1),
-        (KeyCode::Key2, 2),
-        (KeyCode::Key3, 3),
-        (KeyCode::Key4, 4),
-        (KeyCode::Key5, 5),
-        (KeyCode::Key6, 6),
-        (KeyCode::Key7, 7),
-        (KeyCode::Key8, 8),
-        (KeyCode::Key9, 9),
-    ]
-    .iter()
-    {
-        if input.just_pressed(key) {
-            events.send(PowerEvent::Charge(amount));
-        }
-    }
-
-    if input.just_pressed(KeyCode::R) {
-        events.send(PowerEvent::Reserve(1));
-    }
-
-    if input.just_pressed(KeyCode::S) {
-        events.send(PowerEvent::Spend(4));
-    }
-
-    if input.just_pressed(KeyCode::D) {
-        events.send(PowerEvent::Discard(1));
-    }
-
-    if input.just_pressed(KeyCode::A) {
-        events.send(PowerEvent::Add(1));
-    }
-
-    if input.just_pressed(KeyCode::F) {
-        events.send(PowerEvent::Force(1));
-    }
 }
