+++ examples/gaia-project/view/power.rs
@@ -2,18 +2,6 @@ use crate::logic::power::PowerCycleTracker;
 use crate::types::*;
 use bevy::prelude::*;
 
-pub struct PowerViewPlugin;
-
-#[derive(Copy, Clone, Eq, PartialEq)]
-pub enum PowerEvent {
-    Add(u8),
-    Charge(u8),
-    Discard(u8),
-    Force(u8),
-    Reserve(u8),
-    Spend(u8),
-}
-
 pub struct PowerViewState {
     pub tracker: PowerCycleTracker,
 }
